refactor(MoveStrategy): 모든 기물 전략 테스트 위치 수정
- 하나의 파일에 있던 모든 기물 전략 테스트를 각각 분리
